// packages/backend/remotion/types.ts
export type MascotPose =
  | "NORMAL" | "SURPRESO" | "DESCONFIADO" | "BRAVO" | "RINDO" | "IRONICO" | "INDIGNADO"
  | "APONTANDO" | "APONTANDO_ESQUERDA" | "APONTANDO_DIREITA" | "APONTANDO_CIMA" | "APONTANDO_BAIXO"
  | "EXPLICANDO" | "PENSANDO" | "MAO_NA_CABECA" | "DINHEIRO" | "POSITIVO" | "NEGATIVO"
  | "FACEPALM" | "ASSUSTADO" | "DINHEIRO_DOENDO" | "SUSSURRO" | "NEGANDO" | "CONCORDANDO";